# Conflicts:
#	docs/en/migrations.rst
#	src/Phinx/Db/Adapter/AbstractAdapter.php
#	src/Phinx/Db/Table.php
#	tests/Phinx/Db/TableTest.php

+++ src/Phinx/Db/Adapter/TimedOutputAdapter.php
@@ -1,47 +1,26 @@
 <?php
+
 /**
- * Phinx
- *
- * (The MIT license)
- * Copyright (c) 2017 Cake Software Foundation
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated * documentation files (the "Software"), to
- * deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
- * sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
- * IN THE SOFTWARE.
- *
- * @package    Phinx
- * @subpackage Phinx\Db\Adapter
+ * MIT License
+ * For full license information, please view the LICENSE file that was distributed with this source code.
  */
+
 namespace Phinx\Db\Adapter;
 
-use Phinx\Db\Table;
+use BadMethodCallException;
 use Phinx\Db\Table\Column;
 use Phinx\Db\Table\ForeignKey;
 use Phinx\Db\Table\Index;
+use Phinx\Db\Table\Table;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * Wraps any adpter to record the time spend executing its commands
+ * Wraps any adapter to record the time spend executing its commands
  */
-class TimedOutputAdapter extends AdapterWrapper
+class TimedOutputAdapter extends AdapterWrapper implements DirectActionInterface
 {
-
     /**
-     * {@inheritdoc}
+     * @inheritDoc
      */
     public function getAdapterType()
     {
@@ -69,7 +48,8 @@ class TimedOutputAdapter extends AdapterWrapper
      * Write a Phinx command to the output.
      *
      * @param string $command Command Name
-     * @param array  $args    Command Args
+     * @param array $args Command Args
+     *
      * @return void
      */
     public function writeCommand($command, $args = [])
@@ -103,7 +83,7 @@ class TimedOutputAdapter extends AdapterWrapper
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
      *
      * @return void
      */
@@ -116,7 +96,7 @@ class TimedOutputAdapter extends AdapterWrapper
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
      *
      * @return void
      */
@@ -129,40 +109,98 @@ class TimedOutputAdapter extends AdapterWrapper
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @return void
      */
-    public function createTable(Table $table)
+    public function createTable(Table $table, array $columns = [], array $indexes = [])
     {
         $end = $this->startCommandTimer();
         $this->writeCommand('createTable', [$table->getName()]);
-        parent::createTable($table);
+        parent::createTable($table, $columns, $indexes);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
+     */
+    public function changePrimaryKey(Table $table, $newColumns)
+    {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
+        $end = $this->startCommandTimer();
+        $this->writeCommand('changePrimaryKey', [$table->getName()]);
+        $adapter->changePrimaryKey($table, $newColumns);
+        $end();
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
+     */
+    public function changeComment(Table $table, $newComment)
+    {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
+        $end = $this->startCommandTimer();
+        $this->writeCommand('changeComment', [$table->getName()]);
+        $adapter->changeComment($table, $newComment);
+        $end();
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function renameTable($tableName, $newTableName)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand('renameTable', [$tableName, $newTableName]);
-        parent::renameTable($tableName, $newTableName);
+        $adapter->renameTable($tableName, $newTableName);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function dropTable($tableName)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand('dropTable', [$tableName]);
-        parent::dropTable($tableName);
+        $adapter->dropTable($tableName);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @return void
      */
     public function truncateTable($tableName)
     {
@@ -173,113 +211,187 @@ class TimedOutputAdapter extends AdapterWrapper
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function addColumn(Table $table, Column $column)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand(
             'addColumn',
             [
                 $table->getName(),
                 $column->getName(),
-                $column->getType()
+                $column->getType(),
             ]
         );
-        parent::addColumn($table, $column);
+        $adapter->addColumn($table, $column);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function renameColumn($tableName, $columnName, $newColumnName)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand('renameColumn', [$tableName, $columnName, $newColumnName]);
-        parent::renameColumn($tableName, $columnName, $newColumnName);
+        $adapter->renameColumn($tableName, $columnName, $newColumnName);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function changeColumn($tableName, $columnName, Column $newColumn)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand('changeColumn', [$tableName, $columnName, $newColumn->getType()]);
-        parent::changeColumn($tableName, $columnName, $newColumn);
+        $adapter->changeColumn($tableName, $columnName, $newColumn);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function dropColumn($tableName, $columnName)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand('dropColumn', [$tableName, $columnName]);
-        parent::dropColumn($tableName, $columnName);
+        $adapter->dropColumn($tableName, $columnName);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function addIndex(Table $table, Index $index)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand('addIndex', [$table->getName(), $index->getColumns()]);
-        parent::addIndex($table, $index);
+        $adapter->addIndex($table, $index);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function dropIndex($tableName, $columns)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand('dropIndex', [$tableName, $columns]);
-        parent::dropIndex($tableName, $columns);
+        $adapter->dropIndex($tableName, $columns);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function dropIndexByName($tableName, $indexName)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand('dropIndexByName', [$tableName, $indexName]);
-        parent::dropIndexByName($tableName, $indexName);
+        $adapter->dropIndexByName($tableName, $indexName);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function addForeignKey(Table $table, ForeignKey $foreignKey)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand('addForeignKey', [$table->getName(), $foreignKey->getColumns()]);
-        parent::addForeignKey($table, $foreignKey);
+        $adapter->addForeignKey($table, $foreignKey);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return void
      */
     public function dropForeignKey($tableName, $columns, $constraint = null)
     {
+        $adapter = $this->getAdapter();
+        if (!$adapter instanceof DirectActionInterface) {
+            throw new BadMethodCallException('The adapter needs to implement DirectActionInterface');
+        }
         $end = $this->startCommandTimer();
         $this->writeCommand('dropForeignKey', [$tableName, $columns]);
-        parent::dropForeignKey($tableName, $columns, $constraint);
+        $adapter->dropForeignKey($tableName, $columns, $constraint);
         $end();
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @return void
      */
     public function createDatabase($name, $options = [])
     {
@@ -290,7 +402,9 @@ class TimedOutputAdapter extends AdapterWrapper
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @return void
      */
     public function dropDatabase($name)
     {
@@ -301,7 +415,9 @@ class TimedOutputAdapter extends AdapterWrapper
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @return void
      */
     public function createSchema($name = 'public')
     {
@@ -312,13 +428,28 @@ class TimedOutputAdapter extends AdapterWrapper
     }
 
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
+     *
+     * @return void
      */
     public function dropSchema($name)
     {
         $end = $this->startCommandTimer();
         $this->writeCommand('dropSchema', [$name]);
         parent::dropSchema($name);
+        $end();
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return void
+     */
+    public function executeActions(Table $table, array $actions)
+    {
+        $end = $this->startCommandTimer();
+        $this->writeCommand(sprintf('Altering table %s', $table->getName()));
+        parent::executeActions($table, $actions);
         $end();
     }
 }
